@@ -14,38 +14,43 @@
     "url": "git+https://github.com/docsifyjs/docsify.git"
   },
   "main": "lib/docsify.js",
+  "type": "module",
   "unpkg": "lib/docsify.min.js",
   "files": [
     "lib",
     "themes"
   ],
   "scripts": {
-    "bootstrap": "npm i && lerna bootstrap && npm run build:ssr",
-    "serve": "node server",
-    "serve:ssr": "cross-env SSR=1 node server",
+    "bootstrap": "npm i && lerna bootstrap --no-ci",
+    "serve": "npm run _node -- server",
+    "serve:ssr": "cross-env SSR=1 npm run _node -- server",
     "dev": "run-p serve watch:*",
     "dev:ssr": "run-p serve:ssr watch:*",
-    "lint": "eslint .",
-    "fixlint": "eslint . --fix",
-    "test": "jest",
-    "test:e2e": "jest --selectProjects e2e",
-    "test:integration": "jest --selectProjects integration",
-    "test:unit": "jest --selectProjects unit",
-    "css": "node build/css",
+    "lint": "npm run _eslint -- .",
+    "fixlint": "npm run _eslint -- . --fix",
+    "test": "npm run _jest",
+    "test:e2e": "npm run _jest -- --selectProjects e2e",
+    "test:integration": "npm run _jest -- --selectProjects integration",
+    "test:unit": "npm run _jest -- --selectProjects unit",
+    "css": "npm run _node -- build/css.js",
     "watch:css": "npm run css -- -o themes -w",
-    "watch:js": "node build/build.js",
-    "build:css:min": "mkdirp lib/themes && npm run css -- -o lib/themes && node build/mincss.js",
+    "watch:js": "npm run _node -- build/build.js",
+    "build:css:min": "mkdirp lib/themes && npm run css -- -o lib/themes && npm run _node -- build/mincss.js",
     "build:css": "mkdirp themes && npm run css -- -o themes",
-    "build:js": "cross-env NODE_ENV=production node build/build.js",
-    "build:ssr": "node build/ssr.js",
-    "build:cover": "node build/cover.js",
-    "build": "rimraf lib themes && run-s build:js build:css build:css:min build:ssr build:cover",
+    "build:js": "cross-env NODE_ENV=production npm run _node -- build/build.js",
+    "build:cover": "npm run _node -- build/cover.js",
+    "build": "rimraf lib themes && run-s build:js build:css build:css:min build:cover",
     "prepare": "npm run build",
     "pub:next": "cross-env RELEASE_TAG=next sh build/release.sh",
     "pub": "sh build/release.sh",
-    "postinstall": "opencollective-postinstall"
+    "postinstall": "opencollective-postinstall",
+    "_node": "node --experimental-specifier-resolution=node",
+    "_eslint": "DISABLE_V8_COMPILE_CACHE=1 node -- ./node_modules/.bin/eslint",
+    "// ^ _eslint comment": "The DISABLE_V8_COMPILE_CACHE=1 var enables ESM import() in .eslintrc.cjs. See https://github.com/eslint/eslint/issues/13684",
+    "_jest": "npm run _node -- --experimental-vm-modules ./node_modules/.bin/jest",
+    "// ^ _jest comment": "--experimental-vm-modules is needed in order to enable full native ESM support in Jest"
   },
-  "husky": {
+  "husky-OFF": {
     "hooks": {
       "pre-commit": "lint-staged"
     }
@@ -77,11 +82,13 @@
     "copy-dir": "^1.2.0",
     "cross-env": "^6.0.3",
     "cssnano": "^4.1.10",
-    "eslint": "^5.16.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jest": "^24.0.2",
-    "eslint-plugin-jest-playwright": "^0.2.1",
-    "eslint-plugin-prettier": "^3.1.2",
+    "deasync": "^0.1.20",
+    "es-main": "^1.0.2",
+    "eslint": "7.12.1",
+    "eslint-plugin-import": "2.20.1",
+    "eslint-plugin-jest": "24.0.2",
+    "eslint-plugin-jest-playwright": "0.2.1",
+    "eslint-plugin-prettier": "3.1.2",
     "husky": "^3.1.0",
     "jest": "^26.4.2",
     "jest-image-snapshot": "^4.2.0",
@@ -92,6 +99,7 @@
     "mkdirp": "^0.5.1",
     "npm-run-all": "^4.1.5",
     "playwright": "^1.8.0",
+    "playwright-firefox": "^1.5.2",
     "prettier": "^1.19.1",
     "rimraf": "^3.0.0",
     "rollup": "^1.23.1",
