@@ -1,20 +1,23 @@
-const browserSync = require('browser-sync').create();
-const path = require('path');
+import path from 'path';
+import browserSyncModule from 'browser-sync';
+import esMain from 'es-main';
 
+const browserSync = browserSyncModule.create();
 const hasStartArg = process.argv.includes('--start');
+const dirname = path.dirname(import.meta.url.replace('file://', ''));
 
 const serverConfig = {
   host: '127.0.0.1',
   port: 3001,
 };
 
-function startServer(options = {}, cb = Function.prototype) {
+export function startServer(options = {}, cb = Function.prototype) {
   const defaults = {
     ...serverConfig,
     middleware: [
       {
         route: '/_blank.html',
-        handle: function(req, res, next) {
+        handle: function (req, res, next) {
           res.setHeader('Content-Type', 'text/html');
           res.end('');
           next();
@@ -31,18 +34,18 @@ function startServer(options = {}, cb = Function.prototype) {
       },
     ],
     server: {
-      baseDir: path.resolve(__dirname, '../'),
+      baseDir: path.resolve(dirname, '../'),
       routes: {
-        '/docs': path.resolve(__dirname, '../../docs'),
+        '/docs': path.resolve(dirname, '../../docs'),
         '/docs/changelog.md': './CHANGELOG.md',
-        '/lib': path.resolve(__dirname, '../../lib'),
-        '/node_modules': path.resolve(__dirname, '../../node_modules'),
+        '/lib': path.resolve(dirname, '../../lib'),
+        '/node_modules': path.resolve(dirname, '../../node_modules'),
       },
     },
     snippetOptions: {
       rule: {
         match: /<\/body>/i,
-        fn: function(snippet, match) {
+        fn: function (snippet, match) {
           // Override changelog alias to load local changelog (see routes)
           const injectJS = `
             <script>
@@ -75,8 +78,8 @@ function startServer(options = {}, cb = Function.prototype) {
   );
 }
 
-async function startServerAsync() {
-  await new Promise((resolve, reject) => {
+export async function startServerAsync() {
+  await new Promise(resolve => {
     startServer({}, () => {
       console.log('\n');
       resolve();
@@ -84,9 +87,11 @@ async function startServerAsync() {
   });
 }
 
-function stopServer() {
+export function stopServer() {
   browserSync.exit();
 }
+
+const isTopLevelModule = esMain(import.meta);
 
 // Allow starting the test server from the CLI. Useful for viewing test content
 // like fixtures (/index.html)) and local docs site (/docs) used for testing.
@@ -99,13 +104,8 @@ if (hasStartArg) {
   });
 }
 // Display friendly message about manually starting a server instance
-else if (require.main === module) {
-  console.info('Use --start argument to manually start server instance');
+else if (isTopLevelModule) {
+  console.info('Use the --start argument to manually start server instance');
 }
 
-module.exports = {
-  start: startServer,
-  startAsync: startServerAsync,
-  stop: stopServer,
-  TEST_HOST: `http://${serverConfig.host}:${serverConfig.port}`,
-};
+export const TEST_HOST = `http://${serverConfig.host}:${serverConfig.port}`;
