@@ -1,8 +1,8 @@
-const docsifyInit = require('../helpers/docsify-init');
+import docsifyInit from '../helpers/docsify-init';
 
 // Suite
 // -----------------------------------------------------------------------------
-describe('Docs Site', function() {
+describe('Docs Site', function () {
   // Tests
   // ---------------------------------------------------------------------------
   test('coverpage renders and is unchanged', async () => {
