@@ -28,8 +28,9 @@ const compileMedia = {
   },
   iframe(url, title) {
     return {
-      html: `<iframe src="${url}" ${title ||
-        'width=100% height=400'}></iframe>`,
+      html: `<iframe src="${url}" ${
+        title || 'width=100% height=400'
+      }></iframe>`,
     };
   },
   video(url, title) {
@@ -109,6 +110,7 @@ export class Compiler {
         return html;
       })(text);
 
+      // TODO parse() expects an arg, but here it does not receive an arg so it fails.
       const curFileName = this.router.parse().file;
 
       if (isCached) {
@@ -204,7 +206,7 @@ export class Compiler {
      * @param {Number} level Type of heading (h<level> tag)
      * @returns {String} Heading element
      */
-    origin.heading = renderer.heading = function(text, level) {
+    origin.heading = renderer.heading = function (text, level) {
       let { str, config } = getAndRemoveConfig(text);
       const nextToc = { level, title: removeAtag(str) };
 
